Day 16
Ended up not needing the match counter at all. Used it in part 1 hoping that it would be a more modular option in case part 2 required something goofy. Could have been slightly cleaner by just early returning each of those and having a success case at the bottom of the for loop instead.